+++ R/0_utils.R
@@ -17,7 +17,7 @@ brulee_print <- function(x, ...) {
   if (!is.null(x$dims$levels) && !is.null(x$parameters$class_weights)) {
     cat("class weights",
         paste0(
-         lvl,
+          lvl,
           "=",
           format(x$parameters$class_weights),
           collapse = ", "
@@ -77,6 +77,9 @@ cat_schedule <- function(x) {
 
 # ------------------------------------------------------------------------------
 
+#TODO make sure indicies are good with extra result
+# update print method for class weights and val loss
+# cli update
 
 model_to_raw <- function(model) {
   con <- rawConnection(raw(), open = "w")