+++ tests/testthat/test-mlp-multinomial.R
@@ -104,7 +104,7 @@ test_that("basic multinomial mlp SGD", {
               dropout = .1,
               hidden_units = 5,
               optimize = "SGD",
-              batch_size = 64,
+              batch_size = 64L,
               momentum = 0.5,
               learn_rate = 0.1)},
   regex = NA)